requirements.txt
Pillow>=9.0
numpy>=1.23.4